feat: Add ServiceRequest validation and update service creation/edit views
- Implemented ServiceRequest class for comprehensive validation of service requests, including various services like transportation, hotel, documents, handling, and more.
- Updated create and edit views to streamline the form structure and improve user experience.
- Enhanced error messages for better clarity and guidance during form submission.
- Removed unnecessary CSS styles from the edit view to simplify the design.
- Adjusted button styles for action buttons to improve visual hierarchy and user interaction.
